auth_api/python/requirements.txt
Package                Version
---------------------- -------
cffi                   1.15.1 
click                  8.1.3  
cryptography           39.0.1 
Flask                  2.2.2  
importlib-metadata     6.0.0  
itsdangerous           2.1.2  
Jinja2                 3.1.2  
jwt                    1.3.1  
MarkupSafe             2.1.2  
mysql-connector-python 8.0.32 
pip                    19.2.3 
protobuf               3.20.3 
pycparser              2.21   
PyJWT                  2.6.0  
setuptools             41.2.0 
Werkzeug               2.2.2  
zipp                   3.12.1 
Package                Version  
---------------------- ---------
certifi                2022.12.7
cffi                   1.15.1   
charset-normalizer     3.0.1    
click                  8.1.3    
cryptography           39.0.1   
Flask                  2.2.2    
idna                   3.4      
importlib-metadata     6.0.0    
itsdangerous           2.1.2    
Jinja2                 3.1.2    
jwt                    1.3.1    
MarkupSafe             2.1.2    
mysql-connector-python 8.0.32   
pip                    19.2.3   
protobuf               3.20.3   
pycparser              2.21     
PyJWT                  2.6.0    
requests               2.28.2   
setuptools             41.2.0   
urllib3                1.26.14  
Werkzeug               2.2.2    
zipp                   3.12.1   
